ruby: define #to_simple_value on Type instances
It converts to a representation that uses only Hash, Array and plain
strings and numbers. This allows to convert values into

In order to save some performance, it is possible to marshal containers
and arrays of numbers into a packed string using the pack_simple_arrays
option. See the method documentation for more details

Note that /std/string is converted to a String already